+++ src/consensus.cpp
@@ -417,9 +417,6 @@ List split_frequencies(const List trees, const bool exact = false) {
 
 // trees is a list of objects of class phylo, all with the same tip labels
 // (try RenumberTips(trees, trees[[1]]))
-// Per #168, unexpected behaviour if root position differs in non-preorder trees
-// Further investigation could be beneficial; for now, suggest applying
-// the function to preorder trees only.
 // [[Rcpp::export]]
 RawMatrix consensus_tree(const List trees, const NumericVector p,
                          const bool exact = false) {